// sdks/helium-sdk/src/model/boat.ts
/**
 * @helium/ts-tsoa-api
 * Helium API with TSOA generator and openAPI sdk generator
 *
 * The version of the OpenAPI document: 0.0.2
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from '../api';
import { BoatDimensionData } from './boatDimensionData';
import { BoatHardware } from './boatHardware';

export class Boat {
    'id': string;
    'hardware'?: Array<BoatHardware>;
    'dimensionData'?: BoatDimensionData;
    /**
    * The model of the boat
    */
    'model': string;
    /**
    * The brand of the boat
    */
    'brand': string;
    'owner': string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "id",
            "baseName": "_id",
            "type": "string"
        },
        {
            "name": "hardware",
            "baseName": "hardware",
            "type": "Array<BoatHardware>"
        },
        {
            "name": "dimensionData",
            "baseName": "dimensionData",
            "type": "BoatDimensionData"
        },
        {
            "name": "model",
            "baseName": "model",
            "type": "string"
        },
        {
            "name": "brand",
            "baseName": "brand",
            "type": "string"
        },
        {
            "name": "owner",
            "baseName": "owner",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return Boat.attributeTypeMap;
    }
}