// sdks/helium-sdk/src/model/boatHardware.ts
/**
 * @helium/ts-tsoa-api
 * Helium API with TSOA generator and openAPI sdk generator
 *
 * The version of the OpenAPI document: 0.0.2
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from '../api';
import { Part } from './part';
import { System } from './system';

export class BoatHardware {
    'parts'?: Array<Part>;
    'systems'?: Array<System>;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "parts",
            "baseName": "parts",
            "type": "Array<Part>"
        },
        {
            "name": "systems",
            "baseName": "systems",
            "type": "Array<System>"
        }    ];

    static getAttributeTypeMap() {
        return BoatHardware.attributeTypeMap;
    }
}